fix before video

+++ DocumentsPagesViewModels/MaterialsJoined.cs
@@ -4,13 +4,13 @@ namespace DocumentsPagesViewModels
     {
         public string Name { get; set; }
         public string Unity { get; set; }
-        public float Count { get; set; }
-        public float Cost { get; set; } 
-        public float Price { get; set; }
+        public double Count { get; set; }
+        public double Cost { get; set; } 
+        public double Price { get; set; }
         public string VatRate { get; set; }
-        public float Vat { get; set; }
-        public float CostWithVat { get; set; }
-        public float? Weight  { get; set; }
+        public double Vat { get; set; }
+        public double CostWithVat { get; set; }
+        public double? Weight  { get; set; }
         public string Account { get; set; }
     }
 }
